[Chore #139551663] Add test for get maximum priority
